fix: change double to BigDecimal

+++ src/gr/aueb/cf/cf9/ch18/bankapp/dto/AccountReadOnlyDTO.java
@@ -1,5 +1,7 @@
 package gr.aueb.cf.cf9.ch18.bankapp.dto;
 
 
-public record AccountReadOnlyDTO(String iban, double balance) {
+import java.math.BigDecimal;
+
+public record AccountReadOnlyDTO(String iban, BigDecimal balance) {
 }
